requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
py7zr==0.22.0
PyYAML==6.0.2
scikit_learn==1.6.1
seaborn==0.13.2
setuptools==76.0.0
termcolor==3.0.1
timm==1.0.15
torch==2.6.0
torch_geometric==2.6.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
torch-spatiotemporal==0.9.5
lightning==2.1.4
h5py==3.14.0
scikit-survival==0.25.0
lifelines==0.30.0
patsy==1.0.1
SurvivalEVAL==0.4.7
torchsurv==0.1.5
xgboost==3.1.0
pycox==0.3.0